+++ packages/codemod/build.mjs
@@ -2,6 +2,18 @@ import esbuild from "esbuild";
 
 import pkg from "./package.json" with { type: "json" };
 
+import dotenv from "dotenv";
+
+dotenv.config();
+
+const define = {};
+
+for (const k in process.env) {
+  if (k.startsWith("PUBLIC_") === false) continue;
+
+  define[`process.env.${k}`] = JSON.stringify(process.env[k]);
+}
+
 esbuild
   .build({
     entryPoints: ["./src/index.ts"],
@@ -15,6 +27,7 @@ esbuild
     platform: "node",
     target: ["esnext"],
     external: [...Object.keys(pkg.dependencies)],
+    define,
   })
   .catch(() => process.exit(1));
 
@@ -32,5 +45,6 @@ esbuild
     platform: "node",
     target: ["esnext"],
     external: [...Object.keys(pkg.dependencies)],
+    define,
   })
   .catch(() => process.exit(1));